acq path: don't worry if a SECOM confocal has no CCD
Currently, on SECOM, the only point of the OPM is to turn on/off the CCD fan.
It was handling fine already the cases where the CCD has no fan,
but it didn't expect there is no CCD.

=> check for the CCD, and if it's missing but there is a laser-mirror, relax.

So finally, setFan() become setCCDFan(), as it avoids worrying about passing
the component and not calling it when there is no component.