acq path: do not raise LookupError when a stream has no specific optical path configuration
When a stream is passed to setPath() and the OPM doesn't know what do to with it, it
would raise a LookupError. We were supposed to catch such exceptions in the _prepare(),
but it's not that easy because setPath() returns a Future, which raises the exception.
So the exception was never caught, and the stream preparation would actually fail.

=> Don't raise exception in such case, it's just fine to put a log message.