gui cont tabs: handle SPARCv1 + redux in SPARCv2 alignment tab
The major difference in terms of code between the SPARCv1 and SPARCv2 is the mirror stage.
On the SPARCv2, it can park/engage and the alignment position is meaningful only when engaged.

To support SPARCv1 systems with a retro-fitted "redux" mirror stage, make the SPARCv2
alignment tab happy to control a SPARCv1. The optical path manager still knows that it's
a SPARCv1.

The full idea is not yet working, as for now the main GUI still instantiate the SPARCv1 tab in
such a case, so this also need updates.