gui conf: fix get_hw_settings_config() modifying the original data
It was _trying_ to copy the dict to avoid modifying it, but it's actually a recursive
dict, so it needs a recursive copy.
Without this, the standard config could be affected at runtime when calling get_hw_settings_config()
with different roles (including None).

Need some tweaking in the event handler, as a copy modifies the EventBinder.