gui tabs sparc2 align: Replace Momement of Inertia display by a raw CCD display in mirror alignment
Moment of Inertia was a potentially good idea... but in practice it doesn't work
that well. With some type of samples it doesn't work, it has a slow feedback, and
actually using just the raw CCD image is a lot easier than expected.

That means that for the SPARC compact it's not necessary to have 2 modes anymore.
This also brings some support for hybrid SPARCv1/v2 (but the proper tab selection
must still be implemented).