gui cont tabs: let the Tab classes decide themselves if they should be shown and their name
The order, availability, and name of each tab was hard-coded in a declarative way in the main.
That's handy in some cases, but as the decision of which SPARC alignment tab is getting
more complex, it's getting hard to keep using this system. We already relied on a hack
to change the default tab depending on the mirror state.

So instead, let each Tab subclass decide about the default tab and which one is available.

That allows to select the right tab based on which mirror stage is present, instead of
just the role of the microscope.

Also simplify the way the label is set, as it's pretty much static.