acq leech: new Leech concept, with ProbeCurrentAcquirer
New concept in order to perform small acquisition tasks during a long acquisition.
It's mostly for the probe current for now, but hopefully generic enough
to be used in other cases later on.

With some little extention, it should be possible to adapt the drift compensation
to the leech concept. It would even be able to support inter acquisition drift
compensation. But not now yet...