dataio tiff/hdf5: support MD_EBEAM_CURRENT_TIME
The notion of "ebeam current over time" is pretty narrow, and used only
in some rare cases, but when it's measured, we _really_ want to save (and
read) it.

For HDF5, the format is relatively straightforwards, as we can just store
an array of float of shape Nx2.
For OME-TIFF, it's a lot more convoluted. To store a list of tuples, we
need to create a special XML element (Current), and associate to it a
"Time" attribute. At least that was the way that seemed to follow most
the phylosophy of OME XML.

As bonus, we now allow to store MD_EBEAM_CURRENT also on the OME-TIFF.