util dataio: When a greyscale image has no description, don't assume it's Secondary Electrons data
We were defaulting the description to "Secondary Electrons".
Even if it's quire likely, it could well be anything else which just happens to not
have much metadata. Especially, when using an SEM with another detector (eg, Backscatter Electron),
it's pretty confusing.

So instead name it "Electrons", to highlight the fact we don't know much (but it's
probably an EM image).