driver pigcs: For closed-loop stage, report the hardware range as-is
We were trying to be clever, and double the range with the hardware is not
referenced. However, this doesn't help, as anyway the controller will check
that all the moves are within limit. In addition, that caused the range (center)
to change after every start of Odemis, which is confusing for the user.

=> Just report the range as-is.
To deal with the fact the stage is used unreferenced and therefore the 0 might not
be at the center, the controller config (params 0x15 and 0x30) should be doubled.

Note: this causes the range width to be twice less than before. However, in practice
this doesn't limit where the user can go more than before, because the controller
was raising an error (7) in such case. Installations where the stage cannot be
referenced should be updated with piconfig to have a twice larger range, in order
to do what the original code attempted to do (but didn't manage).