acq: new implementation of MeanWeaver
I changed the MeanWeaver, but implemented a different algorithm than we discussed: It builds up the image by 1) inserting the part of the tile where the image is empty, 2) multiplying the tile with a 2d weight matrix with a 1 at the center and 0s at the edges, 3) multiply the region of the current image with 1-weightvector, 4) take the mean. I think this implementation is efficient, easy, doesn't have to keep the images in memory and can insert tiles at any position. Please let me know what you think.