driver tlfw: fix timeout when changing filter
Changing the filter can easily take more than a 1s.
In such case, there won't be input, and it would timeout, causing the move
to report an error. This could also lead the acquisition to start before
the filter is in place.

This ensures that we wait long enough for the move.