gui cont tabs SPARC2 align: fix mirror align mode freezing the GUI
Computing the MoI can be quite CPU intensive. You just cannot do it
only every single image, in the main thread, with large images at
high frame rate, without freezing the GUI.

So lower the expectations: do it maximum once per sec. And also, don't
do it in the GUI thread.

Also fix the SEM spot not being active during the mirror align.
The special MoI stream used to take care of it, but now that it's
a simple CCD stream, we need to use the same trick as for the other
views.