acq algin: Timeout if no data comes from the CCD during autofocusing
Waiting forever can be really long, so it's better to timeout.
Of course, that implies we have _some_ idea of how long it should take,
but we were already checking this for estimating the duration of the
autofocus.

It should help when trying to autofocus on a detector which stops
responding (because it was unplugged, or started overheating...)