gui cont tabs: Fix brightlight not turning on during SPARCv2 autofocus
In commit 1042c7e3f6 ("Replace Momement of Inertia display by a raw CCD display in mirror alignment")
simplified quite a bit the SPARCv2 alignment tab code, but by switching the speclines
stream from BrightfieldStream to CameraStream, it silently dropped the ability for the
stream to turn on the emitter.

Also add/fix some comments related to the alignment tab.