gui cont streams: Fix stream panel appearing in locked_mode even if it shouldn't
Commit 1042c7e3f6 ("Replace Momement of Inertia display by a raw CCD display in mirror alignment")
changed the behaviour of the streambar controller to only show the
stream panels of the focused view if it's in locked_mode.
However, the change was applied on the code running when the view changes.
The decision of showing/hidding a stream panel is also done when adding the stream.
This caused the first time the SPARCv2 alignment tab is displayed to show
two stream panels for the CCD: also the one for the mirror alignment.

=> Check if we are in locked mode when adding a stream too.