gui conf: modify get_local_vas() to get hw_settings from the argument
It used to get the hw_settings by calling get_hw_settings_config().
However, this function takes an optional argument (the role).
If the argument is not passed, it kind-of works, but it's not adapted to
the current microscope. For instance, on the SPARC, it would miss
some local VAs such as the shutter minimum period.

We could pass the role as an argument to get_local_vas() as well, but
as the hw_settings is pretty much static, we just pass this settings.
Now, it's only computed once, and reused everywhere. As by-product,
that should improve very slightly the performance.

Note: it used to work fine until commit 3171c47c8f0fa, because due
to the bug that this commit fixed, the hw_settings were modified at
the beginning, and with some luck, calling get_hw_settings_config()
with role=None would actually be equivalent to calling it with
role="sparc".