gui conf: factorize code to display _any_ value into a user-friendly version into units.value_to_str()
The value of the VA must be converted into a string both at init, when creating
the GUI widget, and whenever the VA value changes.

We already had a code formatting the value when the VA was changing.
It was "optimized" by checking at init what type the value is.
That works well in 99% of the cases, but if the type changes, it goes wrong.
So now, we just check the type every time.

Also use share the code with the display at init. Until now, it could end up
displayed differently for the first value and the next values.