gui comp stream: fix the stream panel "remove" button having "wrong" parent
All the button of the stream panel header should have as parent the header,
the same way as they share the same sizer.
That was not true for the "remove" button. On Linux (GTK), it doesn't seem
to matter, but on Windows, that's probably what prevents the button to be
clickable.