plugins auto_align: Add option to select the reference image
The reference image was hard-coded to the first SEM stream found in the current view.
It kind of work most of the time, however in some use cases it's handy to use
a different image as reference (for instance, an Optical stream).

So we now default to the same SEM stream as before, but also allow the user to
switch afterwards to any "spatial" stream available.