driver andocam2: turn off the cooling on shutdown
It's nice to have the camera cooled down when starting again, but it's not a must.
In other cases, when the camera is water-cooled, it's better to stop cooling
it down as soon as we cannot monitor the whole system, to avoid over-heating.
(Of course the camera has a mechanism to detect over-heating and prevent damages,
but it's better to not rely on it whenever possible)