dataio tiff: Allow reading pyramidal multi-C DataSet and small clean-ups
For now, there is no support for reading tiles on image with more than 2D.
On the other hand, for OME-TIFF, we "merge" optical data with multiple C
(= different wavelengthes) and split it back when reading it. In such case,
it'd technically be considered a 3D array with the C dimension of length 1.
For now, we just "optimize" it and reshape the array into a 2D array.

Also raise the error properly, when it's really a 3D array.