dataio tiff: fix maxzoom of a DataArrayShadow Pyramidal being sometimes incorrect
The maxzoom is infered from the number of subIFD's, which would be checked at
init. When reading such information, the tiff library needs to be told which
directory it is. It worked in the basic cases because we were initialising the
DAS just after reading the TIFF directory. In merged/split DAS, this could fail.