plugins timelapse: don't remove the spot stream from the standard tab
When acquiring we would not take the spot stream. That's fine, but
as the list of steams was not duplicated, it would also remove the
stream from the standard acquisition tab.
As .spotStream is still referencing it, it would still somehow work,
but it's nice to not break things at all.