driver pigcs: Don't try to use Exponent notation when sending numbers
In most cases the numbers would be sent (properly) as in "0.001", but if
the number was getting too small, it would end up sending "12.3e-16".
The PI controller doesn't support such notation, and it would raise error 54.
Actually, this happens quite rarely as the GUI has check to not request tiny
moves (and it's hard to do anyway, with the mouse).

=> Use %f instead of %g