odemisd: detect extra errors in the microscope files
Detect components having duplicate roles.

Detect components missing a class and not having any creator.
It was supposed to be detected, but if the component was "child" of the Microscope,
which almost every component is, odemisd would be optimistic and expect the Microscope
to create it. Of course, the Microscope component has no idea what to do...

Also activate some Errors when validating, which were warning until now.