gui: use merge ratio for history overlay
Using both a built-up overview image and a history overlay at the same time makes the image too cluttered. Therefore we don't want to use the history overlay on the built-up overview image, but only on the global image taken by the camera. The transparency of the overlay changes with the merge ratio.