gui: bugfix and clean-up ovv image with merge slider for history
The history stream (the stream with the rectangles) was not displayed properly. Its size was smaller than the size of the built-up ovv stream. A new function has been created to initialize an ovv image, which makes sure that both images are initialized in the same way. The memory consumption could be improved by changing the size of the history (rectangle) stream to 1x1 pixel. However, when doing this, we get another issue. It seems like the tiles are added to the wrong stream, so sliding to one side gives both the rectangles and the tiles and sliding to the other only a black image. Maybe the streams are sorted somewhere based on their size? Anyway, the size of the image is now reduced to 1200 x 1200 pxs, so initializing to this size instead to 1x1 might not be such a big problem.