acq align delphi: Add I option to move back the sem-stage at its initial position
During twin stage calibration, when focusing the spot, it's easy for the user
to mistakenly move the stage, which forces the sem-stage to the "old" position.
In such case, it's handy to have a way back, instead of restarting the calibration
or looking all around of the spot. => Introduce the "I" key.

Also detect that the blanker is forced on or off, and put it back to automatic.
Also force the ebeam focus to move by a known good absolute value, instead
of depending on the current HFW.