tileacq plugin: pause streams when tiling/ bugfix future
When opening the window, all streams are paused. Before, opening the window caused the opposite stream to play, i.e. if the optical stream was playing, it stopped and the SEM stream started playing when opening the window. The reason for this behaviour is still a mystery. The commit also fixes a bug that occurred when canceling the future before starting the acquisition.