gui/plugins: use acquisition type for icons instead of stream type
acq_type is now added as an argument when intializing a stream. This requires many changes in the controller (every time a stream is initialized). The RGBStream is now not represented in the list of acquisition types (ico_blending_goal icon). I could not find where and what for it was used.