cli: sort axes and dict in general alphabetically
For listing the axes of the actuators, and all the VAs containing dict/set
(which have no order normally), we sort them alphabetically.

This helps a little the reading for axes, as many VAs will share the common
keys (eg, "x", "y", "z"), so it's nice if they are in the same logical order.