driver ueye: Complete reorganisation of the acquisition function
Follow a structure much closer to what the ueyedemo program does.
The uEye lib seems to handle all the thread issues by itself, as well
as most of the reconnections of the camera. The previous code attempted
to also handle the same issues (as it was based on similar code for
other camera libraries which don't handle such corner cases). It appeared
to cause extra issues. In particular, doing a fast stop/start of the DataFlow
would often result in stall of ~15s.

The new code might handle less properly event synchronisation, but for now
there is no user of this functionality in Odemis.

Tested on the Thorlab camera and the IDS camera used for the SPARC compact.
Also tested on a real SPARC compact, with a PMT board connected too.