gui comp viewport: make sure changing view.mpp cannot result in endless loop
VA subscribers should (almost) never set the value of the VA they listen too,
as it's likely to cause infinite loop. Of course, as soon as the value is the same,
the VA detects it, and don't notify the subscribers again.
The _on_view_mpp() subscriber to .mpp would not only do this, but also it's in
a call_in_wx_main(), so its calls are buffered. With the new overview code, the
mpp could be changed twice before this subscriber would be called, which would
result in an endless back and forth change between the two values.

=> Don't see the VA again!