gui: minor fixes
Allow the microscope view to have a larger mpp, as in some corner cases, it's
handy to allow zooming out more. (Mainly when opening TIFF files which have
incorrect/default pixel size).

Also change the stream type of the overview image, to avoid warning that it's incompatible with the view.