config: load plugins from /usr/lib and provide examples in /usr/share
Until now, the plugins were not pacakged, and users had to get them from github.
It's not always very handy.

So now, we ship all the "standard" plugins, in /usr/share, but they are not
loaded.
The plugins which are in /usr/lib are loaded. This structure follows more
what other programs do.
This doesn't affect the plugins which are in ~/.local/share/odemis/plugins,
which are still loaded, as before.