driver pigcs: avoid CL axes staying locked on the physical limits
Detect that the axis is not where the user asked it to go.
In such case, to let the controller try to go there forever,
just accept the current position as the new target position.

The main advantage is that if the user had asked for a position
really far from the limit, it will now be much easier to go
back to within the axis limit.