dataio tiff: fix stiff opening with renamed file
If the original file name cannot be found when opening a serialised TIFF,
we try a little harder by checking if the root file has been renamed, and maybe
the linked file has been renamed similarly.

Also handle a _little_ better if a linked file is missing. It should
mix up less often the metadata.

Ideally, we should create DataArrays based on the OME content, which would
avoid mixing the metadata ever. That's for another day, as it requires a lot
of reorganisation.