acq stream: store live spectrum stream raw data in a standard spectrum shape
Originally, it was thought that this wasn't needed because the live stream
was only used for the "projected" data.
However, when saving a snapshot, the raw data is stored.
To properly store this raw data, we therefore need to put it according
to the standard spectrum shape (= data on the C/5th dim).

This is mostly handy to store background spectrum, directly from the
acquisition tab.