plugins quickcl: ensure the counter increases properly
The counter would be increased, and immediately after, as the filename VA is updated,
it would be set back to its current value.

In most cases, if the previous file was saved, create_filename() would still increase
the counter, as the current filename already exists. However, if some time/date
pattern was also  there it wouldn't work.