driver andorshrk: add support for external component doing led protection
It already supported changing a TTL output (from the spectrograph) when about to turn on a led,
but now also supports letting other components that a led is going to turn on a led.

This allows to do some interlock completely in software.

Also add a slitleds_settle_time argument to control how long to wait before
turning off the detector.