driver andorshrk: don't fail slit move if a protection component fails
If for some reason a protection component fails to change the protection,
or doesn't want it to change, we should still let the slit move go on.

One possible scenario is that the protection component has been disconnected.
It's not handy for the whole system, but at least the spectrograph should keep
working.