acq-stitching: bugfix registrar comparison to right tile
The registrar would return the wrong position when comparing a new tile to the right tile (it was shifted to the right instead of shifting it to the left). This greatly improves the quality of the images acquired in zigzag pattern. Some test images acquired at TU Delft still don't look perfect, but the remaining issues are small in comparison to the previous results. After this fix, all test cases for registration and stitching pass.