acq path: ensure that on SPARC2 the lens-mover is at FAV_POS
Until now the OPM would not touch the lens-mover.
Normally, it doesn't matter as it starts at the FAV_POS, and stays there
all the time. However, if for any reason it's not at the correct position
(eg, user manually references the axis), it should be set back to the
expected position.