gui comp MarkingLineOverlay: round numbers if they are floats
In order to use all the digits of integers the numbers, commit e4fbb12ba93be
force every number to be displayed with all significant numbers.
However, with floats that's just too much.

So, do like the legend and use the range of the data to find out the
significant numbers needed. When the range is unknown, at force back
the rounding if it's a float.