gui cont: ensure that .tab change is called in GUI main thread if it affects the GUI
All functions which modify the GUI must run in the GUI main thread,
otherwise the GUI may crash, withouth warning.
Several subscribers to .tab modify the GUI, and weren't always running in the GUI main thread.

This wasn't noticed, even changing tab happens often because the main reason the tab
changes is when the user presses the tab widget, and this _always_ happens in the GUI
main thread. It could be noticed mostly at the end of an acquisition, when sometimes
(rarely), the GUI would crash when switching to the analysis tab.