acq stream: move .background VA to the base
When now need .background also for some of the LiveStream. There were already
3 other places defining this VA. It's clearer to put it in the base Stream class.

The goal of this VA is to be used in order to subtract an image from the data when displaying it.
Note: the raw data is not modified, so saving it will (only) store the
original data, without the background.

For now, this is to be used on the SPARC alignment tab.