util img DataArray2RGB: map slightly better the whole irange between 0->255
We had found out that using 256/(irange[1] - irange[0]) didn't work
because the largest values would be mapped to 256 ~= 0 (due to overflow).
Using 255.01 avoids it, but then, only the very largest values would be
mapped to 255. Using 255.99 is more accurate.

There is even a slightly better way, but would require an extra addition
on each pixel.

Also fix the test cases to not fail when it's actually fine.