util img: export shouldn't modify the original metadata
In order to simplify the processing, the metadata is sometimes modified before processing.
When this is done, it shouldn't be done on the original metadata.
=> make sure we duplicate the metadata before modifying it.

Among other things, this caused the wavelength of the (pixel) spectrum to be incorrectly displayed
after a spatial view of the spectrum stream has been exported.