gui: reorganise TOOL_* modes
No need to have duplicate values for each tool, one in the model and one in the comp part.
Ideally, indeed they would be separated, and there would be a mapping.
However, this would require a quite a lot of extra code, for no immediate gain.
=> For now, keep it simple by just using the constants of model everywhere.

Also make the tool ready for the FLIM capabilities.