odemis-mic-selector: try harder if EEPROM IDs cannot be found
For no very clear reason the powerhub has sometimes difficulties getting reading
the IDs on the first time. Ideally, the firmware would handle it and try again.
However, it's surprisingly not so easy to do in the firmware, so has a workaround,
we just ask a second time if this happens.

This has been tested with success in AMOLF.