gui util img: don't try to draw a tile of 0 px width
insert_tile_to_image() would always try to rescale and draw the tile,
even if the scaling is so tiny that it'd be 0 px wide (or high).
Not only does it not make much sense, but it also confuses a lot the
underlying rescale function, which would spit errors like:
imgwarp.cpp:1835: error: (-215) dsize.area() || (inv_scale_x > 0 && inv_scale_y > 0) in function resize

=> Detect such a case and don't even try to display the tiny tile.

Also add an extra check in util.img.rescale_hq() so that the
next time it's easier to trace.