plugins: When the AcquisitionDialog is not used anymore, it should be Destroy()'ed
The documentation of wxPython highlights that Dialogs should always end by calling
Destroy(), to free the memory.
In most plugins, this was done when the acquisition would finish properly, but not
in case of cancellation. It could cause memory leakage.

=> Call .Destroy() at the end of the function creating the AcquisitionDialog()

Also fix a bug in timelapse, which didn't automatically pause the streams, as
the other plugins do. This could result in strange acquisition problems.