gui plugin: allow adding streams without showing them on a viewport
Some streams are not visible on the standard viewport.
It's also not handy to have to support every possible type of viewport,
Moreover, if there are many type of streams simultaneously, it's not
useful for the user to see all the viewports.

=> Introduce a way to just no show the image corresponding to a stream.
Instead, it will add just a simple stream panel, and list all these
streams in .hidden_view.