gui cont microscope: Don't entirely fail if the DELPHI focus fails to reference
We now reference the optical focus axis on the DELPHI in order to set the focus
to a good position after loading.
However, if for some reason the referencing fails, it's not a good idea to entirely
stop loading and not allow the user acquire any image.
=> Just log an error and keep going.

Note that the code moving to the "good optical focus" was already checking whether the focus
axis is referenced or not.