driver actuator FixedPositionsActuator: don't log warnings when everything is fine
When moving from one position to another one, some actuators report
the intermediate position. In such case, the position is unlikely to
be supported by the FixedPositionActuator. That's completely normal,
and as long as at the end of the move, the final position is supported,
it's all fine.
=> don't ever listen to position updates while the move we've just
requested is running.

This avoid filling up the log with WARNINGs.