gui cont tabs SECOM acq: fix FluoStream being added on a confocal SECOM
When no optical stream is present, and the user presses the "Optical" button,
the GUI automatically creates a new optical stream. In such a case, on a
confocal SECOM, it would attempt to create a FluoStream, instead of a ConfocalStream.

=> Move the logic for detecting the confocal SECOM so that .onOpticalState()
can use it as well. It now works in a similar way as for the SEM streams.