gui: Create shared settings for confocal streams
All the confocal streams will eventually be acquired simultaneously,
so they share the settings for the laser-mirror (and light).
Until now, we deal with it by just providing the settings as global
settings (ie, directly linked to the hardware settings).

However, this doesn't work anymore with the introduction of the FLIM
streams, which also uses the laser-mirror but with different settings.
That's also the case for the spot mode which will also be introduced.

=> Use a special SettingStream to hold the settings for the confocal
streams.
That means we need to extend the SettingsController to also display
the Stream settings.