gui: Store the component to which the ROIs and spot positions refer
With the support for the confocal microscope, ROI and spot position
can refer to the laser-mirror, instead of the ebeam. So we cannot
hard-code the ebeam everywhere.
=> introduce .fovComp on the tab data model to store this.
For now it's static for the tab. Maybe one day it'll be per view or
dependent on the current stream playing... but there is a lot more
things that need to be updated to handle this. Let's keep it simple for now.

Also move the relative position <-> physical coordinate convertion
functions away from the miccanvas class, as they have nothing to do
with the goal of displaying 2D images.
=> Move some of them to the overlays (as they also contain clipping
and other extra features which we wouldn't want in a generic function),
and some of them to a new util.comp module.