gui win acq: fix visibility icon for non-spatial streams
For the StreamBarController and StreamController to do the right thing,
the StreamBarController should be in the "ignore_view" mode.

Also simplify the adding of the stream, as now .addStream() has all
the arguments we need.

Also remove listening to the local VAs of the streams... because we now
already do it in a more generic way.

Also ensure slightly re-order the event disconnection to avoid
receiving updates while we disconnect.