acq stream RepetitionStream: accept a scanner as optional argument
With an SEM, e-beam is typically both the emitter and the scanner.
However, on a confocal microscope, emitter is the light and the laser-mirror
is the scanner.
A RepetitionStream actually mostly only cares about the scanner, but
we cannot just pass the scanner as the "emitter" as for live mode,
it's useful to have a proper emitter.

Until now we had a workaround in the ScannedTCSettingsStream, but that
prevents to have local VAs on the emitter.