gui comp canvas: on arrow key, move 10% of HFW, instead of always 100px
We used to move by an "absolute" amount of 100px. The move was therefore a little bit
dependent on the HFW, but depending on the complete size of the window, it could have a
large difference. Especially on degenerated cases, it could be dangerous.
For instance, on a window of 1 px (eg, if a canvas is "hidden" but actually gets the focus),
that could mean a very large physical move.

=> Explicitely move by 10% of the HFW. We use the HFW also for Y, in order to have a more
predicatable move size. In the rest of Odemis (and in microscopes in general), the horizontal
size tends to be used over the vertical size.