gui model: put a .tool on the tab models by default
Almost all the tab model did create a .tool VA, several with a single choice TOOL_NONE,
and for the very few which didn't have it there was special code to handle such case.

The change for the FLIM broke these very few tab models (ie, the plugin).
=> Let's simplify and make sure all the tab have a .tool VA.
Especially, it's now easy to change the .choices afterwards.