driver andorshrk: simulate wavelength computation using grating too
Until now, the simulator would report the pixel->wavelength as only a function of
the center wavelength. However, it typically depends on the grating too (the l/mm),
and that's handy to simulate this, so that we can check whether changing the grating
correctly forces an update of the metadata.
=> take into account the l/mm in the computation, even if it's still completely fake values.