acq stream MonochromatorSettingsStream: drop spectrograph argument
The argument was not used, but had been added thinking that one day, maybe,
it could be used in order to add center wavelength and gratting control.
However, since then, we've added control via the GUI. It might still be
improved, to have control of the axes per stream... but most likely it
will involve a more generic way.

=> drop the half-baked functionality

Also update a few comments related.