gui comp viewport/miccanvas: move all fov_hw handling to the viewport
In some cases, the fov_hw would be updated from the miccanvas, while in some
other cases, from the viewport. It worked, but was confusing.
Simplify by moving all the fov_hw management into the viewport.
That means the miccanvas has no idea what the fov_hw is about, and only cares about
the mpp, which is a good thing.

Should have no behaviour change.