acq stream: SettingStream -> ScannerSettingsStream with zoom support
This extends the original SettingStream to support zooming (and roi setting).
This will allow to control the area of acquisition with the FoV in the GUI.

Note that theoretically, the .roi would have been sufficient.
However, using the roi is not very fitting to map the FoV, which means
that a convertion code must go somewhere. In addition, if support for a
ROI in the GUI is also to be implemented, it would be pretty hard to
express both feature simultaneously with just the .roi.

By placing the code here, we can also provide a .resolution to switch
(non-cropped) resolution explicitly, instead of using the scale.