gui: Link confocal zoom to the optical field-of-view
Similarly to the SEM .horizontalFoV which is linked to the EM FoV,
we link the confocal scanner FoV to the optical FoV.
It's not so straightforward as the laser-mirror scanner doesn't provide
a .horizontalFoV as an e-beam, but allow to zoom by only scanning a smaller
area (controlled via the .resolution and .scale).

So we introduce ScannerFoVAdapter, which is in charge of converting
between the typical API expected by the viewport, and the ScannedSettingsStream.

Note that the control is "zoom" instead of HFW. It is possible to show the .horizontalFoV
VA instead of the zoom VA, but it's more complex, and it seems that for users, the
"zoom" control is more intuitive.