driver actuator ConvertStage: fix moveAbs() going to wrong position when called with only one axis
When an absolute move was requested on a single axis (instead of 2),
the other axis position would be considered as 0, instead of using the
_current_ position. This could result in very large and unexcepted move.

In practice, the SPARC and the DELPHI both use this component, but the
GUI never does an absolute move with only axis. So this should not
have happened in a normal usage. It'd still be possible to trigger
such move via the command line. On the SECOM, when using a ConvertStage
to compensate the lens rotation, this would be noticeable.

=> Consider it the current position, as it should.