gui miccanvas: also use .allowed_mode to decide which overlay should be created
The actual tools that might be used on a canvas are defined by the .tools of the
tab data model _and_ the allowed modes of this view.

In particular, this caused the SPARC AR live view to show the ROA and anchor region,
although it does make sense on this view.

=> Compute the possible choices out of both the tools.choices and allowed_mode

The drawback is that we had to give up on the hack to show the ROA on the SECOM
acquisition window. However, that's pretty easy to solve: we just explicitly add
it to the canvas in the acquisition window controller.