driver pigcs: remove incorrect log message
On move update, the log message would be incorrect. It also didn't take
into account the axis inversion.
We could make it correct, but it'd require quite a lot more lines.
As anyway it's possible to check in the log the value also with the return of "MOV?",
which directly reflects the hardware expectation, let's just not try.

Also add a log message in the absolute move similar to the relative move.