gui comp grid: Drop .swap_viewports()
This was an unused and dangerous function as if the visible and invisible
viewports were not updated afterwards, it would leave the object in a bad state.

Also harden get_4_viewports() to not go crazy if the widgets in the grid are not really Viewports.

Replace the test case to test set_visible_viewports(), which is used, instead of swap_viewports().