util runtest: show integration test results for each test
Follow the python.unittest convention, and show the result
of every test. We used to only show error... and only
some of the error were properly logged.

Also always stop the odemis backend between test cases, so that
if a test case went wrong and left the backend running, the next
tests don't get confused.