gui test comp viewport: fix MicroscopeViewportTestCase
After changes in the responsability between the viewport and the miccanvas
on resize and mpp changes, the test case was broken, mainly because it was faking
wx calls which were now not happening in the same way.

=> Just use the standard wx call, to avoid "white-box testing".
Also change the exact checks, to avoid hard-coding values, but
instead compare how the values evolve.