driver tmcm: fix race condition in updatePosition()
The driver supports multiple move simultaneously (as long as they are on different axes).
This means that updatePosition() maybe called simultaneously (with different axes).
As it updates the position by copying it, and then updating the copy, it could
happen that the copy is old by the time it is set again.

=> Protect by a lock as long as we think we have the latest version of the position,
  (and hold the position less long to avoid long locking)

Note: on real hardware it would be very unlikely to happen. On the simulator,
with two moves of the same duration started at the same time, it was reproducable
almost easily.