driver pigcs: Make sure to not move when resuming an closed-loop axis
In a perfect (physical) world, when the axis is suspended, the axis
shouldn't move. However, in our inperfect world, either the axis
drifts, or the encoders report changes in position over time.
So now, we make a fake move to the current position, so that the
target position is updated, and nothing should move.