driver tmcm: fix config extraction failing if the checksum was large
If the checksum (which is pseudo-random) is a large number, the header becomes a large
number, and it would be considered negative. That didn't fit pack() which expected an unsigned int.
We've probably never noticed it before simply because we don't have much variations
in the config stored, and so far we've been lucky.

=> Do like the rest of the pack() and expect an signed int.