driver tmcm: Add timeout to serial write()
Normally the write() never blocks.
However, recently, on some tests, we've noticed it did block.
Might be due to a bug in the kernel (v4.15) or maybe it's just
some specific conditions triggering this block.
In any case, it's never a bad idea to make sure no functions waits forever.